update for export experiment, change the homograph augmentation to support minutiae

+++ superpoint/models/fingernail_minutiae.py
@@ -46,7 +46,9 @@ class FingernailMinutiae(BaseModel):
                                                min_prob=config['detection_threshold'],
                                                keep_top_k=config['top_k']), prob)
             outputs['prob_nms'] = prob
-        pred = tf.to_int32(tf.greater_equal(prob, config['detection_threshold']))
+        # pred = tf.to_int32(tf.greater_equal(prob, config['detection_threshold']))
+        pred = tf.cast(tf.greater_equal(prob, config['detection_threshold']), tf.int32)
+
         outputs['pred'] = pred
 
         return outputs
